requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

openai==1.30.1
python-dotenv==1.0.1

youtube-transcript-api==0.6.2

faiss-cpu==1.7.4

langchain==0.1.20  # Optional if you plan to use LangChain later
jinja2>=3.0.0
pydantic>=2.0.0
fuzzywuzzy
python-Levenshtein
python-multipart